reorganize tests and increase coverage #4230
